refactoring
Isolation of submit method + minors fixs
