+++ src/layout/footer/Footer.tsx
@@ -1,8 +1,8 @@
 import "./footer.css"
 export function Footer() {
-    return (
-        <footer className="footer">
+  return (
+    <footer className="footer">
       <p className="footer-text">Copyright 2020 Argent Bank</p>
     </footer>
-    )
+  )
 }
