+++ src/pages/home/Home.tsx
@@ -1,9 +1,8 @@
-import { features } from 'process';
 import iconChat from '../../assets/img/icon-chat.png';
 import iconMoney from '../../assets/img/icon-money.png';
 import iconSecurity from '../../assets/img/icon-security.png';
 import { FeatureItem } from '../../components/featureItem/FeatureItem';
-import './home.css';
+import './Home.css';
 interface Img {
     src: any,
     alt: string
