+++ src/pages/signIn/SignIn.tsx
@@ -1,23 +1,40 @@
 import { useEffect, useState } from "react";
-import { login } from "../../services/UsersService";
-import "./signIn.css";
+import { isLogged, login } from "../../services/UsersService";
+import "./SignIn.css";
 import { useDispatch, useSelector } from 'react-redux'
 import { setUser } from "../../store/reducers/user/UserReducer";
 import { DateTime } from "luxon";
 import { useNavigate } from "react-router-dom";
-import store from "../../store/store";
 export function SignIn() {
     const [loading, setLoading] = useState(false);
-    const [error, setError] = useState("");
-    const [form, setForm] = useState({ email: "", password: "" });
+    const [form, setForm] = useState({data:{ email: "", password: "" }, error:""});
     const dispatch = useDispatch()
     const navigate = useNavigate()
     const user = useSelector((state: any) => state.user)
-    useEffect(() => {
-        if (user.auth.token !== null &&  DateTime.now().toMillis() < user.auth.expiresAt ) {
-            navigate("/profile")
+    /**
+     * @description handle the form submit, will send the login data to the server and handle the response
+     */
+    const submitForm = async () => {
+        if (isFormValid(form.data).valid) {
+            setLoading(true);
+            const userLogin = await login(form.data.email, form.data.password);
+            setLoading(false)
+            if (userLogin?.body) {
+                dispatch(setUser({...user,auth:{ token: userLogin.body.token, expiresAt: DateTime.now().plus({ days: 1 }).toMillis()} }))
+                navigate("/profile")
+            } else {
+                setForm({...form, error:userLogin?.message})
+            }
+        } else {
+            setForm({...form, error:isFormValid(form.data).error})
         }
-        
+    }
+    useEffect(() => {
+        //handling user loggin state
+        if (isLogged(user.auth.token, user.auth.expiresAt)) {
+            navigate("/profile")
+            return
+        }
     }, [])
     return (
         <main className="main bg-dark">
@@ -28,38 +45,25 @@ export function SignIn() {
                 <form hidden={loading} action="" method="">
                     <div className="input-wrapper">
                         <label htmlFor="username">Username</label>
-                        <input onKeyUp={(e) => setForm({ ...form, email: e.currentTarget.value })} type="text" id="username" />
+                        <input onKeyUp={(e) => setForm({ ...form, data:{...form.data,email: e.currentTarget.value} })} type="text" id="username" />
                     </div>
                     <div className="input-wrapper">
                         <label htmlFor="password">Password</label>
-                        <input onKeyUp={(e) => setForm({ ...form, password: e.currentTarget.value })} type="password" id="password" />
+                        <input onKeyUp={(e) => setForm({ ...form, data:{...form.data,password: e.currentTarget.value} })} type="password" id="password" />
                     </div>
                     <div className="input-remember">
                         <input type="checkbox" id="remember-me" /><label htmlFor="remember-me"
                         >Remember me</label>
                     </div>
-                    <p className="errorSignIn" hidden={error === ""}>{error}</p>
-                    <button onClick={async () => {
-                        if (isFormValid(form).valid) {
-                            setLoading(true);
-                            const userLogin = await login(form.email, form.password);
-                            setLoading(false)
-                            if (userLogin?.body) {
-                                dispatch(setUser({...user,auth:{ token: userLogin.body.token, expiresAt: DateTime.now().plus({ days: 1 }).toMillis()} }))
-                                navigate("/profile")
-                            } else {
-                                setError(userLogin?.message)
-                            }
-                        } else {
-                            setError(isFormValid(form).error);
-                        }
-                    }} type="button" className="sign-in-button">Sign In</button>
+                    <p className="errorSignIn" hidden={form.error === ""}>{form.error}</p>
+                    <button onClick={() =>submitForm()} type="button" className="sign-in-button">Sign In</button>
                 </form>
             </section>
         </main>
 
     )
 }
+
 interface Form {
     email: string;
     password: string;
